internal/<many>: move grpcmetadata reading and writing to it's own package
