+++ internal/cli/base.go
@@ -19,6 +19,7 @@ import (
 	"github.com/hashicorp/waypoint/internal/config"
 	"github.com/hashicorp/waypoint/internal/pkg/flag"
 	pb "github.com/hashicorp/waypoint/internal/server/gen"
+	"github.com/hashicorp/waypoint/internal/server/grpcmetadata"
 )
 
 // baseCommand is embedded in all commands to provide common logic and data.
@@ -304,7 +305,9 @@ func (c *baseCommand) DoApp(ctx context.Context, f func(context.Context, *client
 
 	// Inject the metadata about the client, such as the runner id if it is running
 	// a local runner.
-	ctx = c.project.AddMetadata(ctx)
+	if id, ok := c.project.LocalRunnerId(); ok {
+		ctx = grpcmetadata.AddRunner(ctx, id)
+	}
 
 	// Just a serialize loop for now, one day we'll parallelize.
 	var finalErr error
